+++ examples/plot_fraction_scan.py
@@ -1,7 +1,9 @@
 """Example of fraction scan plot."""
+
 from __future__ import annotations
 
 import numpy as np
+from ftag import get_discriminant
 
 from puma import Line2D, Line2DPlot
 from puma.metrics import calc_eff
@@ -21,7 +23,7 @@ is_b = df["HadronConeExclTruthLabelID"] == 5
 fc_values = np.linspace(0.0, 1.0, 101)
 SIG_EFF = 0.77
 
-dips_scores = df[["dips_pu", "dips_pc", "dips_pb"]].values
+dips_scores = df[["dips_pu", "dips_pc", "dips_pb"]]
 
 
 def calc_effs(fc_value: float):
@@ -37,8 +39,7 @@ def calc_effs(fc_value: float):
     tuple
         Tuple of shape (, 3) containing (fc_value, ujets_eff, cjets_eff)
     """
-    arr = dips_scores
-    disc = arr[:, 2] / (fc_value * arr[:, 1] + (1 - fc_value) * arr[:, 0])
+    disc = get_discriminant(dips_scores, "dips", signal="bjets", fc=fc_value)
     ujets_eff = calc_eff(disc[is_b], disc[is_light], SIG_EFF)
     cjets_eff = calc_eff(disc[is_b], disc[is_c], SIG_EFF)